feat:seperate repsonsabilites for single face  detection
